chore: Rename '/Login' -> '/Authentication'
